multiple/grpc_version/requirements/requirements.txt
grpcio
grpcio-tools
Pillow
numpy<2.0
opencv-python-headless
rembg
onnxruntime
torch
torchvision
torchaudio
requests
rich
matplotlib
